# AppKube-common-security-rbac-services/Api-Specifications/apispecs.yml
openapi: 3.0.3
info:
  title: Common security rbac services
  version: 1.0.0
  description: API for managing organizations, policies, and emails
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /config/byKey/{key}:
    get:
      tags:
        - Config
      summary: Find configuration by key
      operationId: findConfigByKey
      parameters:
        - name: key
          in: path
          description: The key of the configuration to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '404':
          description: Configuration not found
        '500':
          description: Internal server error
  /config/find-by-key-org-id:
    get:
      tags:
        - Config
      summary: Find configuration by key and organization ID
      operationId: findConfigByKeyAndOrgId
      parameters:
        - name: key
          in: query
          description: The key of the configuration to find
          required: true
          schema:
            type: string
        - name: organizationId
          in: query
          description: The ID of the organization
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '404':
          description: Configuration not found
         
        '500':
          description: Internal server error
  /config/find-by-org-id/{organizationId}:
    get:
      tags:
        - Config
      summary: Find configurations by organization ID
      operationId: findConfigsByOrgId
      parameters:
        - name: organizationId
          in: path
          description: The ID of the organization
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Config'
        '404':
          description: No configurations found
         
        '500':
          description: Internal server error
  /config/find-by-created-by/{createdBy}:
    get:
      tags:
        - Config
      summary: Find configurations by creator
      operationId: findConfigsByCreatedBy
      parameters:
        - name: createdBy
          in: path
          description: The username of the creator
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Config'
        '404':
          description: No configurations found
       
        '500':
          description: Internal server error
  /config/find-by-updated-by/{updatedBy}:
    get:
      tags:
        - Config
      summary: Find configurations by updater
      operationId: findConfigsByUpdatedBy
      parameters:
        - name: updatedBy
          in: path
          description: The username of the updater
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Config'
        '404':
          description: No configurations found
         
        '500':
          description: Internal server error
  /config/find-by-status/{status}:
    get:
      tags:
        - Config
      summary: Find configurations by status
      operationId: findConfigsByStatus
      parameters:
        - name: status
          in: path
          description: The status of the configurations
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Config'
        '404':
          description: No configurations found
         
        '500':
          description: Internal server error
  /config/byId/{id}:
    get:
      tags:
        - Config
      summary: Find configuration by ID
      operationId: findConfigById
      parameters:
        - name: id
          in: path
          description: ID of the configuration to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '404':
          description: Configuration not found
        '500':
          description: Internal server error  
  /config/listAll:
    get:
      tags:
        - Config
      summary: Retrieve all configurations
      operationId: findAllConfigs
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Config'
        '500':
          description: Internal server error
  /config/createConfig:
    post:
      tags:
        - Config
      summary: Create a new configuration
      operationId: createConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '400':
          description: Bad request
          
        '500':
          description: Internal server error
  /config/create-encrypted:
    post:
      tags:
        - Config
      summary: Create a new encrypted configuration
      operationId: createEncryptedConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '400':
          description: Bad request
          
        '500':
          description: Internal server error
  /config/configuration/{id}:
    delete:
      tags:
        - Config
      summary: Delete a configuration by ID
      operationId: deleteConfigById
      parameters:
        - name: id
          in: path
          description: ID of the configuration to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Configuration not found
        '500':
          description: Internal server error
...
components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        fax:
          type: string
        dateOfEstablishment:
          type: string
          format: date-time
        status:
          type: string
        cmdbOrgId:
          type: integer
          format: int64
        fileName:
          type: string
        fileStorageLocationType:
          type: string
        profileImage:
          type: string
...
    Config:
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
         
        value:
          type: string
        
        status:
          type: string
         
        isEncrypted:
          type: boolean
         
        organization:
          $ref: '#/components/schemas/Organization'
...